splitting structure into separate files
